+++ CHANGELOG.rst
@@ -3,6 +3,43 @@ Change Log
 
 pykechain changelog
 
+1.11 (UNRELEASED)
+-----------------
+ * In KE-chain 2.5 the way we use task customization has changed drastically. Pykechain (from 1.11 onwards) supports this by implementing a new concept in the activity called `Activity.customization()` (#161). This provides you an `ExtCustomization` object, which you can inspect and add new widgets. Please see the documentation on `ExtCustomization` and `Activity.customization()` for more details. An example to use is:
+
+    >>> activity = project.activity(name='Customizable activity')
+    >>> customization = activity.customization()
+    >>> part_to_show = project.part(name='Bike')
+    >>> customization.add_property_grid_widget(part_to_show, custom_title="My super bike"))
+
+ * removed previously announced deprecated methods for `activity.create_activity()` (use `Activity.create()`) and `property.output` (unused).
+ * Added deprecation warnings when using `InspectorComponent` objects and old style `Customization` components. They will be removed in November 2017 (introduced in pykechain 1.9)
+ * Added the ability to retrieve a list of project members and managers with the `Scope.members()` method (#169)
+ * updated the documentation structure to better access all the pykechain models.
+ * updated tests
+
+1.10.3 (28AUG17)
+----------------
+ * Corrected the creation of partmodels (`Part` with category `MODEL`) with multiplicities other than `ZERO_MANY` as the provided multiplicity option was not respected in the `create_model()` method of `Client` and `Scope`. Thanks @raduiordache for the find. (#170)
+ * Updated tests.
+
+1.10.2 (22AUG17)
+----------------
+ * Corrected the ability to assign multiple assignees, using a list of assignees to an activity using the `Activity.edit()` method. (#167)
+ * Updated tests.
+
+1.10.1 (18AUG17)
+----------------
+ * updated incorrect tests related to `Activity.associated_parts()`. (#96, #149)
+
+1.10.0 (18AUG17)
+----------------
+ * Ability to edit the status of an `Activity`. Please refer to the `ActivityStatus` enumerations. (#163)
+ * Ability to sort properties of a `Part` model. (#141)
+ * Upgraded the requirements of dependent packages for development. (#152, #160, #159, #153, #157, #154)
+ * Added tests for all new features to get the > 95% coverage
+ * Updated the documentation.
+
 1.9.1 (27JUN17)
 ---------------
  * Improved testing. Notably on the new inspector objects. No functional change only that we want to reach our goal of 95% test coverage! Thanks to @raduiordache (#137)